requirements.txt
streamlit==1.38.0
groq==0.9.0
python-dotenv==1.0.1
requests==2.32.3
base64==3.2.1  # For image encoding
pygame==2.6.0  # For audio handling if needed (optional)